breaking: move preds.py -> preds/discovery.py
update imports in scripts and tests
