+++ models/gnome/test_gnome.py
@@ -3,7 +3,7 @@ import pymatviz as pmv
 
 from matbench_discovery.data import Model
 from matbench_discovery.enums import MbdKey
-from matbench_discovery.preds import df_preds
+from matbench_discovery.preds.discovery import df_preds
 
 __author__ = "Janosh Riebesell"
 __date__ = "2024-02-03"
