+++ pyproject.toml
@@ -31,7 +31,9 @@ dependencies = [
   "pandas>=2.2.2",
   "plotly>=5.23",
   "pymatgen>=2024.8.9",
-  "pymatviz[export-figs,df-pdf-export,df-svg-export]>=0.13.2",
+  # "pymatviz[export-figs,df-pdf-export,df-svg-export]>=0.13.2",
+  # temp install from main branch
+  "pymatviz@git+https://github.com/janosh/pymatviz",
   "ruamel.yaml>=0.18.6",
   "scikit-learn>=1.5",
   "scipy>=1.13",
