+++ scripts/analyze_model_failure_cases.py
@@ -20,7 +20,12 @@ from matbench_discovery import PDF_FIGS, SITE_FIGS, WBM_DIR
 from matbench_discovery.data import DataFiles, df_wbm
 from matbench_discovery.enums import MbdKey
 from matbench_discovery.metrics.discovery import classify_stable
-from matbench_discovery.preds import df_each_err, df_each_pred, df_metrics, df_preds
+from matbench_discovery.preds.discovery import (
+    df_each_err,
+    df_each_pred,
+    df_metrics,
+    df_preds,
+)
 
 __author__ = "Janosh Riebesell"
 __date__ = "2023-02-15"
