+++ scripts/calc_model_run_times.py
@@ -18,7 +18,7 @@ from tqdm import tqdm
 
 from matbench_discovery import SITE_FIGS, WANDB_PATH
 from matbench_discovery.data import Model, round_trip_yaml
-from matbench_discovery.preds import model_styles
+from matbench_discovery.preds.discovery import model_styles
 
 __author__ = "Janosh Riebesell"
 __date__ = "2022-11-28"
