+++ scripts/model_figs/rolling_mae_vs_hull_dist_models.py
@@ -11,7 +11,7 @@ from pymatviz.utils import MATPLOTLIB, PLOTLY
 from matbench_discovery import PDF_FIGS, SITE_FIGS
 from matbench_discovery.enums import MbdKey, TestSubset
 from matbench_discovery.plots import rolling_mae_vs_hull_dist
-from matbench_discovery.preds import (
+from matbench_discovery.preds.discovery import (
     df_each_pred,
     df_metrics,
     df_metrics_uniq_protos,
